backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Redis / Celery
redis==5.2.1
celery==5.4.0

# AI
anthropic==0.42.0
google-cloud-vision==3.8.1

# Image processing
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==2.1.3

# Storage (S3-compatible for R2)
boto3==1.35.81

# PDF generation
weasyprint==63.1

# Billing
stripe==11.4.0

# Email
resend==2.5.0

# Utilities
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1

# Monitoring
sentry-sdk[fastapi]==2.19.2
